framework partially set up
