+++ Content/BattleAreas.java
@@ -1,9 +1,3 @@
-import artifact.Artifact;
-import buff.Buff;
-import enemy.Enemy;
-import location.BattleArea;
-import weapon.Verbs;
-import weapon.Weapon;
 import java.util.ArrayList;
 import java.util.List;
