+++ build.gradle
@@ -21,8 +21,8 @@ application {
 
 dependencies {
     implementation group: 'org.json', name: 'json', version: '20220320'
-    implementation group: 'org.fusesource.jansi', name: 'jansi', version: '2.4.0'
     implementation group: 'com.squareup.okhttp3', name: 'mockwebserver', version:'4.10.0'
+    implementation group: 'org.fusesource.jansi', name: 'jansi', version: '2.2.0'
 }
 
 jar {
