+++ src/main/java/Settings/Setting.java
@@ -1,0 +1,13 @@
+package Settings;
+
+public class Setting {
+    private String name;
+
+    public Setting(String name){
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
