+++ src/main/java/tests/Character/CharacterInfo.java
@@ -1,30 +1,30 @@
-package tests.Character;
-
-import character.Character;
-
-public class CharacterInfo {
-
-    public CharacterInfo(){
-        //testing class
-    }
-
-    public void test1StarCharacter(){
-        System.out.println(new Character(1, "test", 0, 0, 0, 0.0, 0, "A test character"));
-    }
-
-    public void test2StarCharacter(){
-        System.out.println(new Character(2, "test", 0, 0, 0, 0.0, 0, "A test character"));
-    }
-
-    public void test3StarCharacter(){
-        System.out.println(new Character(3, "test", 0, 0, 0, 0.0, 0, "A test character"));
-    }
-
-    public void test4StarCharacter(){
-        System.out.println(new Character(4, "test", 0, 0, 0, 0.0, 0, "A test character"));
-    }
-
-    public void test5StarCharacter(){
-        System.out.println(new Character(5, "test", 0, 0, 0, 0.0, 0, "A test character"));
-    }
-}
+//package tests.Character;
+//
+//import character.Character;
+//
+//public class CharacterInfo {
+//
+//    public CharacterInfo(){
+//        //testing class
+//    }
+//
+//    public void test1StarCharacter(){
+//        System.out.println(new Character(1, "test", 0, 0, 0, 0.0, 0, "A test character"));
+//    }
+//
+//    public void test2StarCharacter(){
+//        System.out.println(new Character(2, "test", 0, 0, 0, 0.0, 0, "A test character"));
+//    }
+//
+//    public void test3StarCharacter(){
+//        System.out.println(new Character(3, "test", 0, 0, 0, 0.0, 0, "A test character"));
+//    }
+//
+//    public void test4StarCharacter(){
+//        System.out.println(new Character(4, "test", 0, 0, 0, 0.0, 0, "A test character"));
+//    }
+//
+//    public void test5StarCharacter(){
+//        System.out.println(new Character(5, "test", 0, 0, 0, 0.0, 0, "A test character"));
+//    }
+//}
